dom/messagechannel/tests/test_messageChannel_forceClose.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1176034
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1176034 - start/close</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1176034">Mozilla Bug 1176034</a>
<div id="content"></div>
<pre id="test">
</pre>
  <script type="application/javascript">

  var mc = new MessageChannel();

  try {
    postMessage(42, "*", [ mc.port1, window ]);
    ok(false, "Something went wrong.");
  } catch(e) {
    ok(true, "PostMessage should fail and we should not leak.");
  }

  </script>
</body>
</html>
